requirements.txt
Django>=5.2,<6
djangorestframework>=3.14
drf-yasg
numpy>=1.26
pdf2image 
PyPDF2 
python-docx
scikit-learn
pytesseract
numpy==1.26.4
opencv-python 
pillow
spacy
django-cors-headers
